rag/requirements.txt
# Gemini File Search RAG dependencies
google-genai>=0.3.0
python-dotenv>=1.0.0

# API Server (for Open WebUI integration)
fastapi>=0.104.0
uvicorn>=0.24.0
slowapi>=0.1.9

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
responses>=0.24.0
httpx>=0.25.0
